requirements.txt
kaggle==1.6.17
python-dotenv==1.0.1
notebook==7.3.1
numpy==2.2.0
pandas==2.2.3
scikit-learn==1.6.0
seaborn==0.13.2